Refactor test data.
