+++ modules/portal/test/actions/OidcFrontendActionSpec.scala
@@ -18,7 +18,7 @@ package actions
 
 import cats.effect.IO
 import controllers.UserAccountAccessor
-import org.joda.time.DateTime
+import models.TestApplicationData._
 import org.junit.runner.RunWith
 import org.pac4j.core.profile.CommonProfile
 import org.pac4j.play.scala.Pac4jScalaTemplateHelper
@@ -35,17 +35,6 @@ import scala.concurrent.Future
 
 @RunWith(classOf[JUnitRunner])
 class OidcFrontendActionSpec extends Specification with Mockito with ResultMatchers {
-
-  val frodoUser = User(
-    "fbaggins",
-    "key",
-    "secret",
-    Some("Frodo"),
-    Some("Baggins"),
-    Some("fbaggins@hobbitmail.me"),
-    DateTime.now,
-    "frodo-uuid")
-
   val oidcUsernameField = "username"
 
   val mockUserAccountAccessor: UserAccountAccessor = {
@@ -84,7 +73,7 @@ class OidcFrontendActionSpec extends Specification with Mockito with ResultMatch
       oidcUsernameField)(scala.concurrent.ExecutionContext.global, mockPac4jScalaTemplateHelper)
   }
 
-  val block: UserRequest[AnyContent] => Future[Result] = { request: UserRequest[AnyContent] =>
+  val block: UserRequest[AnyContent] => Future[Result] = { _: UserRequest[AnyContent] =>
     Future.successful(Redirect("/index"))
   }
